requirements.txt
pandas
pyyaml
duckdb